Make united solution file to allow debugging the profiler.
Previously, there were 2 different solutions. It required to run the test app from the command line, making it difficult to debug the attached profiler.
Also, user has to manually run the setup cript under admin.

Now, there is only 1 solution, and the structure of output folders was changed.
With these changes, the profiler dll is not registered in the registry anymore, and admin access therefore not required.
Also, set of the environment variables is refined.
For the .NET Core's Hello World application, the variables are set in the launch json configuration.
User can start the project from Visual Studio, and he will get the profiler attached and will be able to debug it.
For the .NET Framework's HelloWorldFramework application, it looks difficult to set up environment variables to run from Visual Studio. But the required set of variables is saved to RunScript.cmd.
User can double-click the script in the project output directory, and he will get the application running with debugger.

Additionally, setruntimeenv.cmd has been updated to reflect the removal of COM registration and intermediate directory for C++ project has been renamed to obj.