EPMDJ-2315: Support for injection into methods with branching.
Previously, it was not possible to inject anything into methods with if-else, loops, and switch statements.
The reason is that these statements are implemented in .net as branching instructions.
These instructions hold an offset, in bytes, between the branching instruction and the jump destination instruction.
If we naively add an instruction in the middle of such method, these offset will become stale, resulting in invalid
Intermediate Language code.

With these changes, branching instructions are handled differently than other types of instructions.
- Instructions stream now can not only hold instructions, but also allows to add labels in front of an instruction.
- When MethodBody encounters a branching instruction with an offset while decompiling a method, a label is created at
  the target instruction. The original offset is then thrown away, and the label is stored in the branching instruction.
- Then we can safely add instructions we want, the labels remain with their targets.
- When MethodBody compiles, it searches for the target labels and calculates offsets.
  Of course, if the user has not added anything to the method, the offsets should be the same as
  they were before the decompilation.
- User can inject his own branching logic using CreateLabel and MarkLabel methods.