EPMDJ-2315: Allow to inject into methods with try-catch and try-finally clauses.
Each method in .net may have a section, which describes try-catch and try-finally clauses.
These clauses are expressed as byte offsets from beginning of the method to the start of try, catch, or finally statement.
Also length of the statement is stored.

Previously, if we added some instruction in the middle of method with such statements,
the byte offsets and length would become stale, likely resulting into invalid
Intermediate Language code.

Now, MethodBody has special handling for these statements:
- When method is being decompiled, the labels are added to the positions where exception handling statements begin or end.
  The initial offsets and lengths are thrown away then.
- There are ExceptionClause and ExceptionSection objects, which maintain relation between the labels and the error-handling statements.
- When MethodBody is being compiled, the ExceptionClauses recalculate the offsets and lengths. This way, if user has injected
  something, the values will be updated correspondingly. If user has not changed anything, the resulting offsets
  and lengths should be the same as before decompilation.
- Currently there is no support for injecting new try-catch and try-finally clauses, but the object model
  allows to add such ability in the future, if needed.