EPMDJ-2315: Error handling for MethodBody.
The wrong code size passed by InsertSimpleFunction revealed insufficient error handling in the MethodBody constructor.
In Debug builds, compiler initializes the newly allocated blocks of memory with pattern like "fd fd fd fd". There is no such initialization in Release builds.

As a result, the InsertSimpleFunction was working in Debug configuration. The MethodBody saw that CodeSize == 5, and read values from after the end of the input sequence. Fortunately, because of the memory initialization pattern, the bytes "fd fd" did not correspond to any OpCode, and were silently ignored. These 2 mistakes compensated each other, and the test was green.

On the contrary, in the Release builds, the garbage was turned to an instruction. So we had got 4 expected instruction and 1 garbage "instruction", effectively failing the test.

To fix these:
- MethodBody now checks that the incoming std::vector<std::byte> really holds at least CodeSize bytes in the instructions stream part
- MethodBody now throws std::runtime_error if it encounters instruction code not corresponding to any known instruction
- Add tests for these two scenarios.